+++ cibuildwheel/macos.py
@@ -18,13 +18,14 @@ from .options import Options
 from .typing import Literal, PathOrStr, assert_never
 from .util import (
     CIBW_CACHE_PATH,
+    AlreadyBuiltWheelError,
     BuildFrontend,
     BuildSelector,
     NonPlatformWheelError,
     call,
     detect_ci_provider,
     download,
-    find_compatible_abi3_wheel,
+    find_compatible_wheel,
     get_build_verbosity_extra_flags,
     get_pip_version,
     install_certifi_script,
@@ -323,13 +324,13 @@ def build(options: Options, tmp_path: Path) -> None:
                 build_options.build_frontend,
             )
 
-            abi3_wheel = find_compatible_abi3_wheel(built_wheels, config.identifier)
-            if abi3_wheel:
+            compatible_wheel = find_compatible_wheel(built_wheels, config.identifier)
+            if compatible_wheel:
                 log.step_end()
                 print(
-                    f"\nFound previously built wheel {abi3_wheel.name}, that's compatible with {config.identifier}. Skipping build step..."
+                    f"\nFound previously built wheel {compatible_wheel.name}, that's compatible with {config.identifier}. Skipping build step..."
                 )
-                repaired_wheel = abi3_wheel
+                repaired_wheel = compatible_wheel
             else:
                 if build_options.before_build:
                     log.step("Running before_build...")
@@ -409,6 +410,9 @@ def build(options: Options, tmp_path: Path) -> None:
                     shutil.move(str(built_wheel), repaired_wheel_dir)
 
                 repaired_wheel = next(repaired_wheel_dir.glob("*.whl"))
+
+                if repaired_wheel.name in {wheel.name for wheel in built_wheels}:
+                    raise AlreadyBuiltWheelError(repaired_wheel.name)
 
                 log.step_end()
 
@@ -536,7 +540,7 @@ def build(options: Options, tmp_path: Path) -> None:
                     )
 
             # we're all done here; move it to output (overwrite existing)
-            if abi3_wheel is None:
+            if compatible_wheel is None:
                 try:
                     (build_options.output_dir / repaired_wheel.name).unlink()
                 except FileNotFoundError: