+++ cibuildwheel/windows.py
@@ -18,12 +18,13 @@ from .options import Options
 from .typing import PathOrStr, assert_never
 from .util import (
     CIBW_CACHE_PATH,
+    AlreadyBuiltWheelError,
     BuildFrontend,
     BuildSelector,
     NonPlatformWheelError,
     call,
     download,
-    find_compatible_abi3_wheel,
+    find_compatible_wheel,
     get_build_verbosity_extra_flags,
     get_pip_version,
     prepare_command,
@@ -279,13 +280,13 @@ def build(options: Options, tmp_path: Path) -> None:
                 build_options.build_frontend,
             )
 
-            abi3_wheel = find_compatible_abi3_wheel(built_wheels, config.identifier)
-            if abi3_wheel:
+            compatible_wheel = find_compatible_wheel(built_wheels, config.identifier)
+            if compatible_wheel:
                 log.step_end()
                 print(
-                    f"\nFound previously built wheel {abi3_wheel.name}, that's compatible with {config.identifier}. Skipping build step..."
+                    f"\nFound previously built wheel {compatible_wheel.name}, that's compatible with {config.identifier}. Skipping build step..."
                 )
-                repaired_wheel = abi3_wheel
+                repaired_wheel = compatible_wheel
             else:
                 # run the before_build command
                 if build_options.before_build:
@@ -367,6 +368,9 @@ def build(options: Options, tmp_path: Path) -> None:
 
                 repaired_wheel = next(repaired_wheel_dir.glob("*.whl"))
 
+                if repaired_wheel.name in {wheel.name for wheel in built_wheels}:
+                    raise AlreadyBuiltWheelError(repaired_wheel.name)
+
             if build_options.test_command and options.globals.test_selector(config.identifier):
                 log.step("Testing wheel...")
                 # set up a virtual environment to install and test from, to make sure
@@ -420,7 +424,7 @@ def build(options: Options, tmp_path: Path) -> None:
                 shell(test_command_prepared, cwd="c:\\", env=virtualenv_env)
 
             # we're all done here; move it to output (remove if already exists)
-            if abi3_wheel is None:
+            if compatible_wheel is None:
                 shutil.move(str(repaired_wheel), build_options.output_dir)
                 built_wheels.append(build_options.output_dir / repaired_wheel.name)