+++ test/test_manylinuxXXXX_only.py
@@ -53,7 +53,7 @@ project_with_manylinux_symbols = test_projects.new_c_project(
 )
 def test(manylinux_image, tmp_path):
     if utils.platform != "linux":
-        pytest.skip("the docker test is only relevant to the linux build")
+        pytest.skip("the container image test is only relevant to the linux build")
     elif platform.machine() not in ["x86_64", "i686"]:
         if manylinux_image in ["manylinux1", "manylinux2010"]:
             pytest.skip("manylinux1 and 2010 doesn't exist for non-x86 architectures")